Restructured tendermint store
