--- src/store/tendermint/getters.js
@@ -1,5 +1,0 @@
-/**
- * Tendermint getters
- */
-
-export default {};
