+++ src/store/tendermint/index.js
@@ -1,20 +1,50 @@
 /**
- * Tendermint state
+ * Tendermint store
  */
 
-import actions from "./actions";
-import getters from "./getters";
-import mutations from "./mutations";
+import {
+  RpcClient
+} from "tendermint";
+import {
+  WS
+} from "Constants";
 
-export const initialState = {};
+const client = RpcClient(WS);
 
 export default {
   namespaced: true,
-  initialState,
-  state: {
-    ...initialState
-  },
-  getters,
-  actions,
-  mutations
+  state: {},
+  getters: {},
+  mutations: {},
+  actions: {
+    /**
+     * Action to subscribe new block web socket
+     * 
+     * @param {Function} commit
+     */
+    subNewClient({
+      commit,
+      dispatch
+    }) {
+      client.subscribe({
+        query: "tm.event = 'NewBlock'"
+      }, event => {
+        let block = event.block;
+        commit("blocks/addNewBlock", block, {
+          root: true
+        });
+        commit("blocks/setLastBlock", block, {
+          root: true
+        });
+        dispatch("stake/fetchPool", null, {
+          root: true
+        });
+        if (parseInt(block.header.num_txs) > 0) {
+          dispatch("transactions/updateTransactions", block.header.height, {
+            root: true
+          });
+        }
+      });
+    },
+  }
 };
