--- src/store/tendermint/mutations.js
@@ -1,5 +1,0 @@
-/**
- * Tendermint mutations
- */
-
-export default {};
